[windows][netpath] Add windows support for traceroute
DO NOT MERGE

Initial support for traceroute on windows.
As presently implemented may (probably does) break linux build.

Implements the traceroute functionality on windows by making
direct syscalls to the single raw socket.  Uses existing code for packet formatting.
Uses some existing code for packet processing; although it cuts out lots of layers.
